auth.js split into separate actions file
