+++ frontend/src/components/SignInPage/SignInPage.js
@@ -1,6 +1,7 @@
+import React, { Component } from 'react';
 import { Link, Redirect } from 'react-router-dom';
-
-import { auth } from '../actions';
+import { auth } from '../../store/actions';
+import { connect } from 'react-redux';
 
 class Login extends Component {
 	onSubmit = e => {
